+++ ChangeLog.rst
@@ -4,6 +4,7 @@
 
 * Fix #38: Long links wrapping controlled by `--no-wrap-links`.
 * Note: `--no-wrap-links` implies `--reference-links`
+* Fix #87: Decode errors can be handled via command line.
 
 2015.6.21
 =========